Linux: Move ThreadManager to its own header
